requirements.txt
Flask==3.0.3
flask-cors==4.0.1
httpx==0.27.0
cachetools==5.3.3
pycryptodome==3.20.0
protobuf==5.26.1
google-api-python-client==2.128.0